// spacetime-files-ts/example/src/module_bindings/app/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const FileBytes = __t.object("FileBytes", {
  bytes: __t.byteArray(),
  mimeType: __t.string(),
});
export type FileBytes = __Infer<typeof FileBytes>;

export const FileSummary = __t.object("FileSummary", {
  id: __t.u64(),
  path: __t.string(),
  mimeType: __t.string(),
  size: __t.u64(),
  sha256Hex: __t.string(),
  visibility: __t.string(),
  updatedAt: __t.timestamp(),
});
export type FileSummary = __Infer<typeof FileSummary>;

export const Folder = __t.object("Folder", {
  id: __t.u64(),
  ownerUserId: __t.string(),
  path: __t.string(),
  name: __t.string(),
  parentPath: __t.string(),
  createdAt: __t.timestamp(),
  updatedAt: __t.timestamp(),
});
export type Folder = __Infer<typeof Folder>;

export const MyFileSummaries = __t.object("MyFileSummaries", {});
export type MyFileSummaries = __Infer<typeof MyFileSummaries>;

export const MyFolders = __t.object("MyFolders", {});
export type MyFolders = __Infer<typeof MyFolders>;